+++ behavior_machine/library/__init__.py
@@ -4,3 +4,4 @@ from .common_state import IdleState, WaitState
 from .parallel_state import ParallelState
 from .selector_state import SelectorState
 from .common_state import SetFlowState, SetFlowFromBoardState, SaveFlowState
+from .atleastone_state import AtLeastOneState